+++ lib/pleroma/web/mastodon_api/controllers/account_controller.ex
@@ -192,7 +192,7 @@ defmodule Pleroma.Web.MastodonAPI.AccountController do
         :allow_following_move,
         :also_known_as,
         :accepts_chat_messages,
-        :hide_birthday
+        :show_birthday
       ]
       |> Enum.reduce(%{}, fn key, acc ->
         Maps.put_if_present(acc, key, params[key], &{:ok, Params.truthy_param?(&1)})