+++ lib/pleroma/web/router.ex
@@ -449,7 +449,7 @@ defmodule Pleroma.Web.Router do
       post("/accounts/:id/subscribe", AccountController, :subscribe)
       post("/accounts/:id/unsubscribe", AccountController, :unsubscribe)
 
-      get("/birthday_reminders", AccountController, :birthdays)
+      get("/birthdays", AccountController, :birthdays)
     end
 
     post("/accounts/confirmation_resend", AccountController, :confirmation_resend)