+++ priv/repo/migrations/29220116183110_add_birth_date_to_users.exs
@@ -4,7 +4,7 @@ defmodule Pleroma.Repo.Migrations.AddBirthDateToUsers do
   def change do
     alter table(:users) do
       add_if_not_exists(:birthday, :date)
-      add_if_not_exists(:hide_birthday, :boolean, default: false, null: false)
+      add_if_not_exists(:show_birthday, :boolean, default: false, null: false)
     end
   end
 end