+++ CHANGELOG.md
@@ -28,7 +28,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Added `/manifest.json` for progressive web apps.
 - MastoAPI: Support for `birthday` and `show_birthday` field in `/api/v1/accounts/update_credentials`.
 - Configuration: Add `birthday_required` and `birthday_min_age` settings to provide a way to require users to enter their birth date.
-- PleromaAPI: Add `GET /api/v1/pleroma/birthday_reminders` API endpoint
+- PleromaAPI: Add `GET /api/v1/pleroma/birthdays` API endpoint
 
 ### Fixed
 - Subscription(Bell) Notifications: Don't create from Pipeline Ingested replies